requirements.txt
streamlit
groq
pdfplumber
python-docx
pandas
plotly